requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
boto3==1.14.12            # via -r requirements.in
botocore==1.17.12         # via boto3, s3transfer
docutils==0.15.2          # via botocore
jmespath==0.10.0          # via boto3, botocore
python-dateutil==2.8.1    # via -r requirements.in, botocore
s3transfer==0.3.3         # via boto3
six==1.15.0               # via python-dateutil
urllib3==1.25.9           # via botocore
